Fix LumexioApi GET cache TTL type mismatch
env('LUMEXIO_API_CACHE_TTL', 15) reads "15" as a string from .env (env()
only casts true/false/null/empty, never numeric strings), which was
passed straight into now()->addSeconds() — a Carbon\CarbonImmutable
method that requires int|float. This broke every LumexioApi::get() call
at runtime (not just under a hypothetical config:cache, since
var_export would preserve the string type too), a regression introduced
by the GET-caching commit earlier this session. Cast at the config
boundary so any consumer gets an int.

Discovered while implementing Task 3 (Screens\Profile) — Profile::mount()
calls LumexioApi::get('/auth/me'), which hit this immediately, and the
full suite was already red (284/317) before any Task 3 changes.

+++ config/lumexio.php
@@ -9,5 +9,5 @@ return [
      * made. Balances snappy tab-switching against data freshness; kept short
      * since dashboard/stock figures change frequently.
      */
-    'cache_ttl' => env('LUMEXIO_API_CACHE_TTL', 15),
+    'cache_ttl' => (int) env('LUMEXIO_API_CACHE_TTL', 15),
 ];
